added package.json, moved server logic into index.js so it can be used as a library
- server logic now in index.js so spacebrew can be used as a library through spacebrew.createServer
- added package.json so spacebrew can be used as a dependency
